+++ week2/src/assets/검정 어글리슈즈.png
