+++ week2/src/assets/자스 deepdive.jpeg
