+++ week2/src/assets/체크셔츠.jpeg
