+++ week2/src/assets/흰색 반팔.jpeg
