+++ week2/src/js/storage.js
