+++ week2/src/js/utils/domHelper.js
